deleted fish list component from fish.js and added new folder
