+++ scripts/Fish.js
@@ -10,15 +10,3 @@ export const Fish = (fish) => {
     </section>
     `
 }
-export const FishList = () => {
-    // Get a reference to the `<article class="content">` element
-    const contentElement = document.querySelector("insert selector here")
-    const fishes = useFish()
-
-    // Add to the existing HTML in the content element
-    contentElement.innerHTML += `
-        <article class="fishList">
-            All the fish go here!
-        </article>
-    `
-}
