+++ scripts/main.js
@@ -1,5 +1,6 @@
 console.log("Welcome to the main module")
 
+import { FishList } from './Fish.js'
 import { useFish } from './FishDataProvider.js'
 
 const allTheFish = useFish()
@@ -7,3 +8,5 @@ const allTheFish = useFish()
 for (const fish of allTheFish) {
     console.log(fish)
 }
+// Import the FishList here
+FishList()
